# swagger_server/models/picture.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from models.base_model_ import Model
import util


class Picture(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, value: str=None):  # noqa: E501
        """Picture - a model defined in Swagger

        :param value: The value of this Picture.  # noqa: E501
        :type value: str
        """
        self.swagger_types = {
            'value': str
        }

        self.attribute_map = {
            'value': 'value'
        }

        self._value = value

    @classmethod
    def from_dict(cls, dikt) -> 'Picture':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Picture of this Picture.  # noqa: E501
        :rtype: Picture
        """
        return util.deserialize_model(dikt, cls)

    @property
    def value(self) -> str:
        """Gets the value of this Picture.


        :return: The value of this Picture.
        :rtype: str
        """
        return self._value

    @value.setter
    def value(self, value: str):
        """Sets the value of this Picture.


        :param value: The value of this Picture.
        :type value: str
        """

        self._value = value