# swagger_server/models/user_token.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from models.base_model_ import Model
import util


class UserToken(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, token: str=None, expiration: str=None):  # noqa: E501
        """UserToken - a model defined in Swagger

        :param token: The token of this UserToken.  # noqa: E501
        :type token: str
        :param expiration: The expiration of this UserToken.  # noqa: E501
        :type expiration: str
        """
        self.swagger_types = {
            'token': str
        }

        self.attribute_map = {
            'token': 'token'
        }

        self._token = token

    @classmethod
    def from_dict(cls, dikt) -> 'UserToken':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserToken of this UserToken.  # noqa: E501
        :rtype: UserToken
        """
        return util.deserialize_model(dikt, cls)

    @property
    def token(self) -> str:
        """Gets the token of this UserToken.


        :return: The token of this UserToken.
        :rtype: str
        """
        return self._token

    @token.setter
    def token(self, token: str):
        """Sets the token of this UserToken.


        :param token: The token of this UserToken.
        :type token: str
        """

        self._token = token