# swagger_server/models/user_update.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from models.base_model_ import Model
import util


class UserUpdate(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, newusername: str=None):  # noqa: E501
        """UserUpdate - a model defined in Swagger

        :param newusername: The newusername of this UserUpdate.  # noqa: E501
        :type newusername: str
        """
        self.swagger_types = {
            'newusername': str
        }

        self.attribute_map = {
            'newusername': 'newusername'
        }

        self._newusername = newusername

    @classmethod
    def from_dict(cls, dikt) -> 'UserUpdate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserUpdate of this UserUpdate.  # noqa: E501
        :rtype: UserUpdate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def newusername(self) -> str:
        """Gets the newusername of this UserUpdate.


        :return: The newusername of this UserUpdate.
        :rtype: str
        """
        return self._newusername

    @newusername.setter
    def newusername(self, newusername: str):
        """Sets the newusername of this UserUpdate.


        :param newusername: The newusername of this UserUpdate.
        :type newusername: str
        """

        self._newusername = newusername